+++ src/java/DTO/CTPN_DTO.java
@@ -4,12 +4,12 @@ public class CTPN_DTO {
     private int maPN;
     private int maSach;
     private int soLuong;
-    private float donGia;
+    private int donGia;
     
     public CTPN_DTO() {
     }
 
-    public CTPN_DTO(int maPN, int maSach, int soLuong, float donGia) {
+    public CTPN_DTO(int maPN, int maSach, int soLuong, int donGia) {
         this.maPN = maPN;
         this.maSach = maSach;
         this.soLuong = soLuong;
@@ -40,15 +40,15 @@ public class CTPN_DTO {
         this.soLuong = soLuong;
     }
 
-    public float getDonGia() {
+    public int getDonGia() {
         return donGia;
     }
 
-    public void setDonGia(float donGia) {
+    public void setDonGia(int donGia) {
         this.donGia = donGia;
     }
 
-    public float calculate(int soLuong, float donGia ){
+    public int calculate(int soLuong, int donGia ){
         return soLuong * donGia;
     }
 }
